// discord-bot/src/classes/slash-commands/interfaces/command/command.interface.ts
import {
  AutocompleteInteraction,
  ChatInputCommandInteraction,
} from "discord.js";

import { CustomClient } from "../../../client/client.class";
import Category from "../../../../enums/category.enum";

export interface ICommand {
  client: CustomClient;
  name: string;
  description: string;
  category: Category;
  options: object;
  default_member_permissions: bigint;
  dm_permission: boolean;

  execute(interaction: ChatInputCommandInteraction): void;
  autoComplete(interaction: AutocompleteInteraction): void;
}